feat: take composer configuration from extra section
Before this change, the configuration for this library was taken from the config section within composer.json. This was invalid as the `config` section belongs to configuration from Composer itself. Moving it to the `extra` section which is intended for this purpose (composer docs: "arbitrary extra data for consumption by scripts")